Friendlier PR comment, helps with #105
https://github.com/cla-assistant/cla-assistant/issues/105#issuecomment-194215536

If you haven't signed the CLA for a project, it's very likely your first
PR ever to this project. And the first thing that happens, within
seconds, is a comment by this bot with an imperative to jump through
another hoop? Oops, I think we'd like to at least show appreciation for
their effort.