Fix race for out-of-order webhooks.
It is possible that the `opened` action may not be the first webhook actually
received for a PR. If this should occur, users would see the CLA check for a PR
appear to hang until a new commit is pushed to the PR.

This change breaks the filtering and handling of a request into two
separately-exported functions, `accepts` and `handle`. Only acceptable
requests will be enqueued for debouncing. As a convenience to operators, the
no-op case will result in a 204 response code. I have confirmed that GitHub
treats a 204 as a successful response code in a private instance of CLA
checker.

Fixes #545