improve request performance for getSignedCLA() (#739)
On an instance with a lot of configured repos (like cla-assistant.io) the current implementation gets a big list of repos first and then performs a second request to determine for which of these the user has a signature on file

This implementation reverses the request order and first requests all CLAs the user has signed and then performs the lookup of repos/orgs to check if that is still the up-to-date CLA. Generally this performs way better as a lot less data needs to be transferred. In my tests I saw an improvement down from a dozen seconds to less than a second.